+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 
 ### Bug fixes
 
+- [#170][GL-170] - Enchantment and attribute modifier text no longer hidden in scenario view, edit and vote menus on Paper servers for Minecraft 1.20.5+
 - [#169][GL-169] - NPE when activating the Randomized Crafts or Upside Down Crafting scenario on Paper servers for Minecraft 1.20.5+
 
+[GL-170]: https://gitlab.com/uhccore/uhccore/-/issues/170
 [GL-169]: https://gitlab.com/uhccore/uhccore/-/issues/169
 
 ## [1.20.10] - 2024-06-26